struggle

+++ src/game/board/show.rs
@@ -4,13 +4,13 @@ use super::Board;
 
 
 impl<'a> Board<'a> {
-    pub fn print(&self, act : Option<i32>) {
+    pub fn print(&self, act : Option<u8>) {
         println!("[T = {:4<}]           {}", self.turn, Unit::title_1());
         println!("                  {}", Unit::title_2());
         for (i, unit) in self.units.iter().enumerate() {
             let ac = match act {
                 None => " ",
-                Some(a) => if i as i32 == a {">"} else {" "},
+                Some(a) => if i as u8 == a {">"} else {" "},
             };
             println!("{ac}{i} : {}{} {} {}", unit.txt_state(), unit.txt_bound(), unit.txt_attr(), unit.txt_state_time());
         }
